Abstract base command functionality to util
